servidor_HTTP/requirements-dev.txt
fastapi
uvicorn[standard]
mysql-connector-python
black
isort
httpx
mock
coverage
pytest-cov